🐛 TT resize (#1071)
- Resize TT on `ucinewgame` after Hash is changed
- Report TT memory size correctly when <8 MB (debug mode/logs)
- Fix `TranspositionTableSize` value in `appsettings.json` (it was a string)

+++ src/Lynx/Model/TranspositionTable.cs
@@ -64,7 +64,7 @@ public static class TranspositionTableExtensions
         {
             ttLength = BitOperations.RoundUpToPowerOf2(ttLength) >> 1;    // / 2
         }
-        var ttLengthMb = ttLength / 1024 / 1024;
+        var ttLengthMb = (double)ttLength / 1024 / 1024;
 
         if (ttLength > int.MaxValue)
         {
